(Restructure) For Landing and Dashboard Page
- Restructure folder to 2 Main folder
  a. landing
  b. dashboard
