+++ src/App.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react';
 
-import Login from './pages/Auth/Login';
-import Contact from './pages/Contact';
-import Home from './pages/Home';
+import Login from './landing/pages/auth/Login';
+import Contact from './landing/pages/general/Contact';
+import Home from './landing/pages/general/Home';
 
 // import nothinghere from './components/nothinghere';
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
-import About from './pages/About';
-import Formregister from './components/Formregister';
+import About from './landing/pages/general/About';
+import Formregister from './landing/components/Formregister';
 // import Formregisterhook from './components/Hook';
-import TnCPage from './pages/T&C';
+import TnCPage from './landing/pages/general/T&C';
 // import '../semantic/dist/semantic.min.css';
 
 class App extends Component {
