--- src/components/successRegisnotif.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Message } from 'semantic-ui-react'
-
-const successRegisnotif = () => (
-  <Message
-    success
-    header='Welcome to Young Innovators Fellowship, Next Leader!'
-    content='Kami akan memberi update jika kamu terpilih ke tahap selanjutnya.'
-  />
-)
-
-export default successRegisnotif;
