+++ src/landing/components/regis.js
