+++ src/landing/config/index.js
